ContactInfoForm: use form data when possible to generate choices for state
When generating choices for the state fields, check if the form data has
country fields; if so, use them. This is needed as we're modifying the
choices for state dynamically; our instantiation-time logic to determine
state choices is insufficient.

This is needed if we want django to validate the state fields properly
(being a ChoiceField, it checks that a field's value is indeed one of
the choices).