ContactInfoForm: don't assume that user has selected a country
Check if the form data submitted contains a valid country before passing
it to area_choices_for_country().

Also, don't bother checking if IN_COUNTRY_ONLY is on; the country field
isn't rendered in the form.

A small change to the precedence of 'copy_address' is squashed in too;
previously, it took effect only if the user didn't select a shipping
country. Change this to take effect always.